+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Automatically minify internal export/import names for esm and system output ([#2087](https://github.com/rollup/rollup/pull/2087))
 * Improve deprecation handling ([#2076](https://github.com/rollup/rollup/pull/2076))
 * Enable `--perf` timings in watch mode ([#2065](https://github.com/rollup/rollup/pull/2065))
+* Improve error handling for plugins ([#2100](https://github.com/rollup/rollup/pull/2100))
 * Remove duplicate badges from readme ([#2083](https://github.com/rollup/rollup/pull/2083))
 * Update readme examples ([#2086](https://github.com/rollup/rollup/pull/2086))